Move DataStandardError to classes/exceptions.py
(with the provision that the directory structure is gonna change later)
