+++ plasmapy/classes/exceptions.py
@@ -1,0 +1,6 @@
+from plasmapy.utils import PlasmaPyError
+
+
+class DataStandardError(PlasmaPyError):
+    """An exception for when HDF5 is not defined by OpenPMD standard."""
+    pass
