+++ plasmapy/utils/__init__.py
@@ -6,7 +6,6 @@ from .checks import (check_quantity,
 from .exceptions import (PlasmaPyError,
                          PhysicsError,
                          RelativityError,
-                         DataStandardError,
                          PlasmaPyWarning,
                          PhysicsWarning,
                          CouplingWarning,
