+++ plasmapy/utils/exceptions.py
@@ -29,11 +29,6 @@ class RelativityError(PhysicsError):
     pass
 
 
-class DataStandardError(PlasmaPyError):
-    """An exception for when HDF5 is not defined by OpenPMD standard."""
-    pass
-
-
 # ----------
 # Warnings:
 # ----------
